+++ src/tr1/game/objects/general/puzzle_hole.c
@@ -115,5 +115,5 @@ void PuzzleHole_Collision(int16_t item_num, ITEM *lara_item, COLL_INFO *coll)
         return;
     }
 
-    Inv_Display(INV_KEYS_MODE);
+    Inv_DisplayKeys(item->object_id);
 }