+++ src/tr1/game/objects/traps/midas_touch.c
@@ -109,5 +109,5 @@ void MidasTouch_Collision(int16_t item_num, ITEM *lara_item, COLL_INFO *coll)
         return;
     }
 
-    Inv_Display(INV_KEYS_MODE);
+    Inv_DisplayKeys(item->object_id);
 }